+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------
 
 - Improve moshi-sealed KSP error messages.
+- Fix fallback adapter support in IR code gen not recognizing Moshi parameters to primary constructors.
 
 0.27.0
 ------